+++ main.go
@@ -69,10 +69,13 @@ func registerHandlers(ctx context.Context) {
 	discovery_kit_sdk.Register(extrabbitmq.NewRabbitVhostDiscovery(ctx))
 	discovery_kit_sdk.Register(extrabbitmq.NewRabbitNodeDiscovery(ctx))
 	discovery_kit_sdk.Register(extrabbitmq.NewRabbitQueueDiscovery(ctx))
+	discovery_kit_sdk.Register(extrabbitmq.NewRabbitExchangeDiscovery(ctx))
 
 	// Actions
 	action_kit_sdk.RegisterAction(extrabbitmq.NewPublishRabbitFixedAmount())
 	action_kit_sdk.RegisterAction(extrabbitmq.NewPublishRabbitPeriodically())
+	action_kit_sdk.RegisterAction(extrabbitmq.NewPublishExchangeFixedAmount())
+	action_kit_sdk.RegisterAction(extrabbitmq.NewPublishExchangePeriodically())
 	action_kit_sdk.RegisterAction(extrabbitmq.NewAlterQueueMaxLengthAttack())
 
 	// Checks